Fix baml_language cyclic dependencies
- Moved shared testing utilities into `baml_project::testing` behind a feature flag
- Removed some dead dependencies
