+++ baml_language/crates/baml_tests/Cargo.toml
@@ -20,7 +20,7 @@ baml_compiler2_tir = { workspace = true }
 baml_compiler_diagnostics = { workspace = true }
 baml_db = { workspace = true }
 baml_fmt = { workspace = true }
-baml_project = { workspace = true }
+baml_project = { workspace = true, features = [ "testing" ] }
 baml_type = { workspace = true }
 baml_workspace = { workspace = true }
 bex_engine = { workspace = true }
