+++ baml_language/crates/bex_engine/Cargo.toml
@@ -36,7 +36,7 @@ wasm-bindgen-futures = { workspace = true }
 
 [dev-dependencies]
 baml_builtins2 = { workspace = true }
-baml_tests = { workspace = true }
+baml_project = { workspace = true, features = [ "testing" ] }
 sys_llm = { workspace = true }
 sys_native = { workspace = true, features = [ "bundle-http" ] }
 anyhow = { workspace = true }
