+++ baml_language/crates/bex_engine/tests/common/mod.rs
@@ -9,7 +9,7 @@
 use std::{io::Write, sync::Arc};
 
 use baml_builtins2::{PromptAst as BuiltinPromptAst, PromptAstSimple};
-use baml_tests::engine::compile_source;
+use baml_project::testing::compile_source;
 use bex_engine::{BexEngine, BexExternalValue, FunctionCallContextBuilder};
 use bex_external_types::BexExternalAdt;
 use bex_vm_types::Program;
