+++ baml_language/crates/bex_vm/Cargo.toml
@@ -35,8 +35,7 @@ web-time = { workspace = true }
 baml_builtins2_codegen = { workspace = true }
 
 [dev-dependencies]
-baml_compiler2_emit = { workspace = true }
-baml_tests = { workspace = true }
+baml_project = { workspace = true, features = [ "testing" ] }
 anyhow = { workspace = true }
 divan = { workspace = true }
 indexmap = { workspace = true }
