+++ baml_language/crates/bex_vm/tests/early_yield.rs
@@ -20,7 +20,7 @@ use std::sync::{
     atomic::{AtomicBool, Ordering},
 };
 
-use baml_tests::engine::compile_source;
+use baml_project::testing::compile_source;
 use bex_vm::{BexVm, VmExecState};
 use bex_vm_types::Value;
